% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincounterR.R
\name{bincounterR}
\alias{bincounterR}
\title{Count Events in Bins}
\usage{
bincounterR(
  x,
  y,
  nbins = c(5, 5),
  Ranges = matrix(c(-Inf, Inf, -Inf, Inf), 2, 2)
)
}
\arguments{
\item{x}{a matrix of numbers or a list}

\item{y}{a matrix of numbers}

\item{nbins}{=c(5,5) number of bins in x and in y direction}

\item{Ranges}{=matrix(c(-Inf, Inf, -Inf, Inf),2,2) a 2x2 matrix with lower and upper bounds}
}
\value{
a list counts and vals
}
\description{
This function discretizes continuous data in two dimensions
}
\keyword{internal}
