% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.R
\name{case.studies}
\alias{case.studies}
\title{Generate Case Studies}
\usage{
case.studies(
  which,
  n = 200,
  nx = n,
  ny = n,
  nbins = -1,
  Ranges = matrix(c(-Inf, Inf, -Inf, Inf), 2, 2),
  ReturnCaseNames = FALSE
)
}
\arguments{
\item{which}{name of the case study.}

\item{n}{=500, sample size.}

\item{nx}{=n, sample size of x data set.}

\item{ny}{=n, sample size of y data set.}

\item{nbins}{=-1,number of bins if discrete data is wanted.}

\item{Ranges}{=matrix(c(-Inf, Inf, -Inf, Inf),2,2) a 2x2 matrix with lower and upper bounds}

\item{ReturnCaseNames}{=FALSE should only list of names be returned?}
}
\value{
a list of functions
}
\description{
This function creates the functions needed to run the various case studies.
}
\keyword{internal}
