% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.marginals.R
\name{change.marginals}
\alias{change.marginals}
\title{Change Marginals of 2D Data}
\usage{
change.marginals(dta, which, theta)
}
\arguments{
\item{dta}{a list with two data sets}

\item{which}{method for modifying the data sets.}

\item{theta}{how much to modify}
}
\value{
a list of functions
}
\description{
This function modifies the marginals of some data set
}
\keyword{internal}
