% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq2D_test_cont.R
\name{chisq2D_test_cont}
\alias{chisq2D_test_cont}
\title{This function does the chi square two sample test for continuous data in two dimensions}
\usage{
chisq2D_test_cont(
  dta_x,
  dta_y,
  Ranges = matrix(c(-Inf, Inf, -Inf, Inf), 2, 2),
  nbins = c(5, 5),
  minexpcount = 5
)
}
\arguments{
\item{dta_x}{a matrix of numbers}

\item{dta_y}{a matrix of numbers}

\item{Ranges}{=matrix(c(-Inf, Inf, -Inf, Inf),2,2) a 2x2 matrix with lower and upper bounds}

\item{nbins}{=c(5,5) number of bins in x and y direction}

\item{minexpcount}{=5 minimum counts required per bin}
}
\value{
a list with statistics and p values
}
\description{
This function does the chi square two sample test for continuous data in two dimensions
}
