% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq2D_test_disc.R
\name{chisq2D_test_disc}
\alias{chisq2D_test_disc}
\title{This function does the chi square two sample test for continuous data in two dimensions}
\usage{
chisq2D_test_disc(dta, minexpcount = 5)
}
\arguments{
\item{dta}{a list with values and counts}

\item{minexpcount}{=5 minimum counts required per bin}
}
\value{
a list with statistics and p values
}
\description{
This function does the chi square two sample test for continuous data in two dimensions
}
