% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqTS.R
\name{chiTS.cont}
\alias{chiTS.cont}
\title{Chi-square test for continuous data, example of user supplied function}
\usage{
chiTS.cont(x, y, TSextra)
}
\arguments{
\item{x}{a matrix of numbers}

\item{y}{a matrix of numbers}

\item{TSextra}{list with some info}
}
\value{
either a test statistic or a p value
}
\description{
This function does the chi square two sample test for continuous data in two dimensions
}
