% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqTS.R
\name{chiTS.disc}
\alias{chiTS.disc}
\title{Chi-square test for discrete data, example of user supplied function}
\usage{
chiTS.disc(x, y, vals_x, vals_y, TSextra)
}
\arguments{
\item{x}{a vector with counts}

\item{y}{a vector with counts}

\item{vals_x}{a vector with points}

\item{vals_y}{a vector with points}

\item{TSextra}{list with some info}
}
\value{
either a test statistic or a p value
}
\description{
This function does the chi square two sample test for discrete data in two dimensions
}
