% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_dist}
\alias{find_dist}
\title{find pairwise Euclidean distances}
\usage{
find_dist(dta)
}
\arguments{
\item{dta}{A list of matrices}
}
\value{
A  matrix
}
\description{
find pairwise Euclidean distances
}
\keyword{internal}
