% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_cont}
\alias{gen_cont}
\title{simulate continuous data}
\usage{
gen_cont(x, y, TSextra)
}
\arguments{
\item{x}{first data set}

\item{y}{second data set}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate continuous data
}
\keyword{internal}
