% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_disc}
\alias{gen_disc}
\title{simulate new discrete data}
\usage{
gen_disc(x, y, vals_x, vals_y, TSextra)
}
\arguments{
\item{x}{first data set, counts}

\item{y}{second data set, counts}

\item{vals_x}{values of discrete random variable x}

\item{vals_y}{values of discrete random variable y}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate new discrete data
}
\keyword{internal}
