% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Order}
\alias{Order}
\title{equivalent to R command order}
\usage{
Order(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
integer vector
}
\description{
equivalent to R command order
}
\keyword{internal}
