% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rank}
\alias{Rank}
\title{equivalent to R command rank}
\usage{
Rank(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
integer vector
}
\description{
equivalent to R command rank
}
\keyword{internal}
