% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDalitz.R
\name{rDalitz}
\alias{rDalitz}
\title{This function creates data for Dalitz plot}
\usage{
rDalitz(nyadd = 0, which = 1, addon, n = 200, nx = n, ny = n)
}
\arguments{
\item{nyadd}{=0}

\item{which}{type of Daliz plot.}

\item{addon}{what to add to uniform}

\item{n}{=200, sample size of both data sets.}

\item{nx}{=n, sample size of x data set.}

\item{ny}{=n, sample size of y data set.}
}
\value{
a list of functions
}
\description{
This function creates data for Dalitz plot
}
\keyword{internal}
