% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosample_test_adjusted_pvalue.R
\name{simpvals}
\alias{simpvals}
\title{Helper function to find test statistics of simulated data.}
\usage{
simpvals(
  dta,
  TS,
  typeTS,
  TSextra,
  A,
  Continuous,
  Ranges,
  nbins,
  minexpcount,
  B
)
}
\arguments{
\item{dta}{a list}

\item{TS}{test statistic routine}

\item{typeTS}{type of routine}

\item{TSextra}{a list}

\item{A}{a matrix}

\item{Continuous}{logical}

\item{Ranges}{a matrix}

\item{nbins}{a vector}

\item{minexpcount}{an integer}

\item{B}{number of simulation runs}
}
\value{
a matrix
}
\description{
Helper function to find test statistics of simulated data.
}
\keyword{internal}
