% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methods.R
\name{test_methods}
\alias{test_methods}
\title{This function checks whether the correct methods have been requested}
\usage{
test_methods(doMethods, Continuous, ReturnMethodNames = FALSE)
}
\arguments{
\item{doMethods}{="all" Which methods should be included?}

\item{Continuous}{is data continuous}

\item{ReturnMethodNames}{= FALSE should names of methods be returned?}
}
\value{
TRUE/FALSE or a character vector
}
\description{
This function checks whether the correct methods have been requested
}
\keyword{internal}
