% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transform01}
\alias{transform01}
\title{transform data to interval (0,1).}
\usage{
transform01(dta)
}
\arguments{
\item{dta}{a list.}
}
\value{
a list
}
\description{
transform data to interval (0,1).
}
\keyword{internal}
