% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont}
\alias{TS_cont}
\title{Find test statistics for continuous data}
\usage{
TS_cont(x, y, TSextra)
}
\arguments{
\item{x}{A matrix.}

\item{y}{A matrix}

\item{TSextra}{list with some helper functions}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for continuous data
}
\keyword{internal}
