% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_cont_pval.R
\name{TS_cont_pval}
\alias{TS_cont_pval}
\title{This function runs a number of two sample tests which find their own p value.}
\usage{
TS_cont_pval(x, y)
}
\arguments{
\item{x}{a matrix of numbers if data is continuous or of counts  if data is discrete.}

\item{y}{a matrix of numbers if data is continuous or of counts  if data is discrete.}
}
\value{
A list of two numeric vectors, the test statistics and the p values.
}
\description{
This function runs a number of two sample tests which find their own p value.
}
