% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_disc}
\alias{TS_disc}
\title{Find test statistics for discrete data}
\usage{
TS_disc(x, y, vx, vy, TSextra)
}
\arguments{
\item{x}{a vector of integers.}

\item{y}{a vector of integers.}

\item{vx}{a vector of real.}

\item{vy}{a vector of real.}

\item{TSextra}{list with extra infos}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for discrete data
}
\keyword{internal}
