% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.R
\name{case.studies}
\alias{case.studies}
\title{Create case studies}
\usage{
case.studies(
  which,
  n = 200,
  nx = n,
  ny = n,
  nbins = -1,
  Ranges = matrix(c(-Inf, Inf, -Inf, Inf), 2, 2),
  ReturnCaseNames = FALSE
)
}
\arguments{
\item{which}{name of the case study.}

\item{n}{=200, sample size for both data sets}

\item{nx}{=n, sample size for first data sets}

\item{ny}{=n, sample size for second data sets}

\item{nbins}{=-1, number of bins for chi-square test, 2D only}

\item{Ranges}{= matrix(c(-Inf, Inf, -Inf, Inf), 2, 2), ranges of variables}

\item{ReturnCaseNames}{=FALSE, should list of case studies be returned?}
}
\value{
a list of functions and vectors
}
\description{
This function creates the functions needed to run the various case studies.
}
